Revert "Cleanup"
This reverts commit 2248776d12c6598f485e7326909c021101eef02a.

+++ src/include/vm/bytecode_handlers.h
@@ -459,12 +459,12 @@ VM_OP_HOT void OpInitBool(tpl::sql::BoolVal *result, bool input) {
   result->val = input;
 }
 
-VM_OP_HOT void OpInitInteger(tpl::sql::Integer *result, int64_t input) {
+VM_OP_HOT void OpInitInteger(tpl::sql::Integer *result, int32_t input) {
   result->is_null = false;
   result->val = input;
 }
 
-VM_OP_HOT void OpInitReal(tpl::sql::Real *result, double input) {
+VM_OP_HOT void OpInitReal(tpl::sql::Real *result, float input) {
   result->is_null = false;
   result->val = input;
 }
